+++ src/pages/popup/Popup.tsx
@@ -20,7 +20,7 @@ Popup.Title = () => {
   );
 };
 
-Popup.Login = ({ isLoaded, userEmail, onLoginWithGoogle }: LoginProps) => {
+Popup.Login = ({ isLoaded, userEmail }: LoginProps) => {
   if (isLoaded) {
     return (
       <LoadingStyle>
@@ -35,7 +35,7 @@ Popup.Login = ({ isLoaded, userEmail, onLoginWithGoogle }: LoginProps) => {
       {userEmail === '' && (
         <LoginStyle>
           <span>로그인이 되어있지 않습니다. 로그인을 해주세요!</span>
-          <GoogleLoginButton onLoginWithGoogle={onLoginWithGoogle} />
+          <GoogleLoginButton />
         </LoginStyle>
       )}
       {userEmail === '' || (
