+++ src/pages/popup/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { auth } from '../../firebase';
-import { GoogleAuthProvider, signInWithCredential } from 'firebase/auth';
 import { ThemeProvider } from 'styled-components';
 import { theme } from '../../styles/theme';
 import GlobalStyles from '../../styles/global';
@@ -23,38 +22,12 @@ function PopupLayout() {
     <Popup>
       <Popup.Content>
         <Popup.Title />
-        <Popup.Login
-          isLoaded={isLoaded}
-          userEmail={userEmail}
-          onLoginWithGoogle={getGoogleAuthCredential}
-        />
+        <Popup.Login isLoaded={isLoaded} userEmail={userEmail} />
       </Popup.Content>
       <Popup.Footer />
     </Popup>
   );
 }
-
-const getGoogleAuthCredential = () => {
-  chrome.identity.getAuthToken({ interactive: true }, token => {
-    if (chrome.runtime.lastError) {
-      console.error(
-        '[Pro-Solve] Token을 받아오던 중 문제가 발생했습니다! :>> ',
-        chrome.runtime.lastError.message,
-      );
-    }
-
-    const credential = GoogleAuthProvider.credential(null, token);
-    signInWithCredential(auth, credential)
-      .then(firebaseAuth => {
-        console.log('[Pro-Solve] Firebase 사용자 인증 정보 :>>', firebaseAuth);
-      })
-      .catch(error => {
-        if (error.code === 'auth/invalid-credential') {
-          chrome.identity.removeCachedAuthToken({ token }, getGoogleAuthCredential);
-        }
-      });
-  });
-};
 
 const root = document.createElement('div');
 document.body.appendChild(root);
