+++ src/types/popup.d.ts
@@ -1,11 +1,6 @@
 interface LoginProps {
   isLoaded: boolean;
   userEmail: string;
-  onLoginWithGoogle: () => void;
 }
 
-interface GoogleLoginProps {
-  onLoginWithGoogle: () => void;
-}
-
-export { LoginProps, GoogleLoginProps };
+export { LoginProps };
